+++ src/components/Cards/ICardsInfoProps.ts
@@ -1,0 +1,4 @@
+export interface ICardsInfoProps {
+  title: string;
+  info: string | string[];
+}
