+++ src/components/Games/Cards.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { Box, Button } from "@mui/material";
-import { Container, Typography, Chip } from "@mui/material";
 import { useTheme } from "@mui/material/styles";
 import { TransitionGroup, CSSTransition } from "react-transition-group";
 import ClearIcon from "@mui/icons-material/Clear";
 import EditIcon from "@mui/icons-material/Edit";
 import { useCardsContext } from "../context/GamesCardsContextProvider";
-import { ICardsInfoProps } from "../AI/Cards/AICardComponent";
 import { CardsInfo } from "../Cards/CardsInfo";
+import { CardInfoMix } from "../Cards/CardsInfoMix";
+import { ModalCardBox } from "../Cards/ModalCardBox";
 
 export interface Card {
   id: string;
@@ -16,64 +16,6 @@ export interface Card {
   price: number;
   currency: string;
 }
-
-export const CardInfoMix: React.FC<ICardsInfoProps> = ({ title, info }) => {
-  const theme = useTheme();
-  return (
-    <Container
-      sx={{
-        display: "flex",
-        alignItems: "baseline",
-        gap: 2,
-        flexDirection: "row",
-        marginBottom: 2,
-        paddingLeft: { xs: 1, sm: 1, md: 1, lg: 2 },
-      }}
-    >
-      <Typography
-        variant="h4"
-        color="text.primary"
-        sx={{
-          width: { xs: "5em", sm: "5em", md: "5em", lg: "4.5em" },
-          textAlign: { xs: "left", sm: "left", md: "left", lg: "right" },
-          flexShrink: 0,
-        }}
-      >
-        {title}
-      </Typography>
-      <Box
-        sx={{
-          display: "flex",
-          flexWrap: "wrap",
-          justifyContent: "center",
-          gap: 1,
-        }}
-      >
-        {Array.isArray(info) ? (
-          info.map((option, index) => (
-            <Chip
-              key={index}
-              label={option}
-              sx={{
-                margin: theme.spacing(0, 0.5),
-                display: "flex",
-              }}
-            />
-          ))
-        ) : (
-          <Chip
-            label={info}
-            sx={{
-              margin: theme.spacing(0, 0.5),
-              padding: { xs: 0.5, sm: 1, md: 1.5 },
-              //TODO: Add hyphenation
-            }}
-          />
-        )}
-      </Box>
-    </Container>
-  );
-};
 
 export const CardComponent: React.FC<{ card: Card }> = ({ card }) => {
   const theme = useTheme();
@@ -191,50 +133,15 @@ const CardsList: React.FC = () => {
 };
 
 export const CardComponentMix: React.FC<{ card: Card }> = ({ card }) => {
-  const theme = useTheme();
-
   return (
-    <Box
-      sx={{
-        flexDirection: "column",
-        paddingTop: 2,
-        paddingBottom: 2,
-        paddingLeft: { md: 0, lg: 1, xl: 2 },
-        paddingRight: { md: 0, lg: 0 },
-        maxWidth: { md: "95%", lg: "95%" },
-        border: "2px solid",
-        borderColor: theme.palette.text.primary,
-        py: 2.5,
-        borderRadius: 2,
-        position: "relative",
-        marginBottom: 5,
-        bgcolor: "background.paper",
-        "&::after": {
-          content: '""',
-          position: "absolute",
-          bottom: theme.spacing(1),
-          left: {
-            xs: theme.spacing(1),
-            sm: theme.spacing(2),
-            md: theme.spacing(4),
-          },
-          right: {
-            xs: theme.spacing(0.7),
-            sm: theme.spacing(1),
-            md: theme.spacing(2),
-          },
-          height: "2px",
-          backgroundColor: theme.palette.text.primary,
-        },
-      }}
-    >
+    <ModalCardBox>
       <CardInfoMix title="Name of the Game" info={card.name} />
       <CardInfoMix title="Difficulty" info={card.difficulty} />
       <CardInfoMix
         title="Price"
         info={card.price.toString() + " " + card.currency}
       />
-    </Box>
+    </ModalCardBox>
   );
 };
